# garage/backend/app/Http/Controllers/Auth/AuthController.php
<?php

namespace App\Http\Controllers\Auth;

use App\Enums\BookingStatus;
use App\Http\Controllers\Controller;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Auth;
use App\Models\User;
use App\Models\Driver;
use Illuminate\Support\Facades\Hash;
use Illuminate\Validation\ValidationException;
use App\Http\Controllers\ParserController;
use App\Models\City;
use App\Models\DriverSpecification;
use App\Models\DriverDoc;
use App\Models\Division;
use App\Models\RentTerm;
use Session;
use Illuminate\Support\Facades\Http;
use Illuminate\Support\Facades\Storage;
use App\Enums\UserStatus;
use App\Enums\UserType;
use App\Enums\CarClass;
use App\Enums\TransmissionType;
use App\Enums\ReferralStatus;
use App\Enums\FuelType;
use App\Models\Referral;
use App\Models\Manager;
use App\Models\Park;
use App\Services\RewardService;
use Illuminate\Support\Carbon;
use Illuminate\Support\Str;

class AuthController extends Controller
{


    /**
     * Получение данных пользователя (аутентифицированный запрос)
     *
     * @OA\Get(
     *     path="api/user",
     *     operationId="GetUser",
     *     summary="Получение данных пользователя (аутентифицированный запрос)",
     *     tags={"Authentication"},
     *     security={{"bearerAuth": {}}},
     * @OA\Response(
     *     response=200,
     *     description="Успешная аутентификация или регистрация",
     *     @OA\JsonContent(
     *         @OA\Property(
     *             property="user",
     *             type="object",
     *             description="Данные пользователя",
     *             @OA\Property(property="user_status", type="string", description="Статус пользователя"),
     *             @OA\Property(property="phone", type="string", description="Номер телефона пользователя"),
     *             @OA\Property(property="name", type="string", nullable=true, description="Имя пользователя"),
     *             @OA\Property(property="email", type="string", nullable=true, description="Email пользователя"),
     *             @OA\Property(property="user_type", type="string", description="Тип пользователя"),
     *             @OA\Property(property="city_name", type="string", description="Название города"),
     *             @OA\Property(
     *                 property="referral_info",
     *                 type="object",
     *                 description="Данные по реферральной программе",
     *                     @OA\Property(property="referral_code", type="string", description="код для рефферальной ссылки"),
     *                     @OA\Property(property="status", type="string", description="Статус реферральной программы" ,ref="#/components/schemas/ReferralStatus"),
     *                     @OA\Property(property="coins", type="integer", description="Количесство заработанных монет")
     *             ),
     *             @OA\Property(
     *                 property="docs",
     *                 type="array",
     *                 description="Данные документов водителя",
     *                 @OA\Items(
     *                     type="object",
     *                     @OA\Property(property="type", type="string", description="Тип документа"),
     *                     @OA\Property(property="url", type="string", nullable=true, description="URL документа")
     *                 )
     *             ),
     *             @OA\Property(
     *                 property="bookings",
     *                 type="array",
     *                 description="Список бронирований",
     *                 @OA\Items(
     *                     type="object",
     *                     @OA\Property(property="id", type="integer", description="Идентификатор бронирования"),
     *                     @OA\Property(property="status", type="string", description="Статус бронирования",ref="#/components/schemas/BookingStatus"),
     *                     @OA\Property(property="start_date", type="string", description="Дата начала бронирования в формате 'd.m.Y H:i'"),
     *                     @OA\Property(property="end_date", type="string", description="Дата окончания бронирования в формате 'd.m.Y H:i'"),
     *                     @OA\Property(
     *                         property="rent_term",
     *                         type="object",
     *                         description="Условия аренды",
     *                         @OA\Property(property="deposit_amount_daily", type="number"),
     *                         @OA\Property(property="deposit_amount_total", type="number"),
     *                         @OA\Property(property="minimum_period_days", type="integer"),
     *                         @OA\Property(property="is_buyout_possible", type="integer"),
     *                         @OA\Property(
     *                             property="schemas",
     *                             type="array",
     *                             @OA\Items(
     *                                 type="object",
     *                                 @OA\Property(property="daily_amount", type="integer"),
     *                                 @OA\Property(property="non_working_days", type="integer"),
     *                                 @OA\Property(property="working_days", type="integer")
     *                             )
     *                         )
     *                     ),
     *                     @OA\Property(
     *                         property="car",
     *                         type="object",
     *                         description="Информация об автомобиле",
     *                         @OA\Property(property="id", type="integer"),
     *                         @OA\Property(property="fuel_type", type="string", description="Тип топлива",ref="#/components/schemas/FuelType"),
     *                         @OA\Property(property="transmission_type", type="string", description="Тип трансмиссии",ref="#/components/schemas/TransmissionType"),
     *                         @OA\Property(property="сar_class", type="string", description="Класс тарифа", ref="#/components/schemas/CarClass"),
     *                         @OA\Property(property="brand", type="string"),
     *                         @OA\Property(property="model", type="string"),
     *                         @OA\Property(property="year_produced", type="integer"),
     *                         @OA\Property(property="images", type="array", @OA\Items(type="string"), description="Ссылки на изображения"),
     *                         @OA\Property(
     *                             property="division",
     *                             type="object",
     *                             description="Информация о дивизионе",
     *                             @OA\Property(property="address", type="string"),
     *                             @OA\Property(property="coords", type="string"),
     *                             @OA\Property(property="phone", type="string"),
     *             @OA\Property(
     *                 property="working_hours",
     *                 type="array",
     *                 description="Расписание работы парка",
     *                 @OA\Items(
     *                     type="object",
     *                     @OA\Property(property="day", type="string", description="День недели на английском",ref="#/components/schemas/DayOfWeek"),
     *                     @OA\Property(
     *                         property="start",
     *                         type="object",
     *                         description="Время начала работы",
     *                         @OA\Property(property="hours", type="integer", description="Часы (0-23)"),
     *                         @OA\Property(property="minutes", type="integer", description="Минуты (0-59)")
     *                     ),
     *                     @OA\Property(
     *                         property="end",
     *                         type="object",
     *                         description="Время окончания работы",
     *                         @OA\Property(property="hours", type="integer", description="Часы (0-23)"),
     *                         @OA\Property(property="minutes", type="integer", description="Минуты (0-59)")
     *                     )
     *                 )
     *             ),
     *                             @OA\Property(
     *                                 property="park",
     *                                 type="object",
     *                                 description="Информация о парке",
     *                                 @OA\Property(property="url", type="string"),
     *                                 @OA\Property(property="commission", type="integer"),
     *                                 @OA\Property(property="park_name", type="string"),
     *                                 @OA\Property(property="about", type="string"),
     *                             )
     *                         )
     *                     )
     *                 )
     *             )
     *         )
     *     )
     * ),
     *     @OA\Response(
     *         response=422,
     *         description="Ошибка валидации",
     *         @OA\JsonContent(
     *             @OA\Property(
     *                 property="errors",
     *                 type="object",
     *                 description="Список ошибок валидации"
     *             )
     *         )
     *     )
     * )
     *
     * @param \Illuminate\Http\Request $request
     * @return \Illuminate\Http\JsonResponse
     */
    public function GetUser(Request $request)
    {
        $user = Auth::guard('sanctum')->user();
        $user->user_type = UserType::from($user->user_type)->name;
        $user->user_status = UserStatus::from($user->user_status)->name;
        $driver = Driver::where('user_id', $user->id)->with('city')->first();
        $referral = Referral::where('user_id', $user->id)->select('referral_code', 'status', 'coins')->first();
        $driverDocs = DriverDoc::where('driver_id', $driver->id)->first(['image_licence_front', 'image_licence_back', 'image_pasport_front', 'image_pasport_address', 'image_fase_and_pasport']);
        $docs = [];
        foreach ($driverDocs->toArray() as $key => $value) {
            $docs[] = [
                'type' => $key,
                'url' => $value  ? asset('uploads') . '/' . $value : null,
            ];
        }
        if (!$driver->city) {
            $user->city_name = 'Москва';
        } else {
            $user->city_name = $driver->city->name;
        }
        $bookings =  $driver->booking;
        if ($bookings) {
            foreach ($bookings as $booking) {
                $booking->status = BookingStatus::from($booking->status)->name;
                $booking->start_date = Carbon::parse($booking->booked_at)->toIso8601ZuluString();
                $booking->end_date = Carbon::parse($booking->booked_until)->toIso8601ZuluString();
                $booking->car->сar_class = CarClass::from($booking->car->tariff->class)->name;
                $booking->car->transmission_type = TransmissionType::from($booking->car->transmission_type)->name;
                $booking->car->fuel_type = FuelType::from($booking->car->fuel_type)->name;
                $booking->car->images = json_decode($booking->car->images);
                $booking->car->division = Division::where('id', $booking->car->division_id)->with('park')->select('address', 'park_id', 'coords', 'working_hours', 'phone')->first();
                $booking->rent_term = RentTerm::where('id', $booking->car->rent_term_id)
                    ->with(['schemas' => function ($query) use ($booking) {
                        $query->where('id', $booking->schema_id);
                    }])
                    ->select('deposit_amount_daily', 'deposit_amount_total', 'minimum_period_days', 'is_buyout_possible', 'id')
                    ->first();
                $workingHours = json_decode($booking->car->division->working_hours, true);
                $booking->car->division->working_hours = $workingHours;
                unset(
                    $booking->created_at,
                    $booking->updated_at,
                    $booking->booked_at,
                    $booking->booked_until,
                    $booking->park_id,
                    $booking->driver_id,
                    $booking->car->booking,
                    $booking->car->created_at,
                    $booking->car->updated_at,
                    $booking->car->status,
                    $booking->car->car_id,
                    $booking->car->division_id,
                    $booking->car->park_id,
                    $booking->car->tariff_id,
                    $booking->car->tariff,
                    $booking->car->rent_term_id,
                    $booking->car->division->id,
                    $booking->car->division->park_id,
                    $booking->car->division->city_id,
                    $booking->car->division->created_at,
                    $booking->car->division->updated_at,
                    $booking->car->division->name,
                    $booking->car->division->park->id,
                    $booking->car->division->park->id,
                    $booking->car->division->park->API_key,
                    $booking->car->division->park->created_at,
                    $booking->car->division->park->updated_at,
                    $booking->car->division->park->created_at,
                    $booking->rent_term->id
                );
                foreach ($booking->rent_term->schemas as $schema) {
                    unset($schema->created_at, $schema->updated_at, $schema->id, $schema->rent_term_id);
                }
            }
        }

        $user->referral_info = $referral;
        $user->docs = $docs;
        $user->bookings = $bookings ? $bookings : null;
        unset($user->id, $user->code, $user->role_id, $user->avatar, $user->email_verified_at, $user->settings, $user->created_at, $user->updated_at);
        return response()->json(['user' => $user]);
    }
    /**
     * Аутентификация пользователя или регистрация нового
     *
     * @OA\Post(
     *     path="api/user/login",
     *     operationId="loginOrRegister",
     *     summary="Аутентификация пользователя или регистрация нового",
     *     tags={"Authentication"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="phone", type="string", example="1234567890", description="Номер телефона пользователя"),
     *             @OA\Property(property="code", type="integer",  description="Код аутентификации"),
     *             @OA\Property(property="type", type="string",  description="Тип пользователя", ref="#/components/schemas/UserType"),
     *             @OA\Property(property="api_key", type="string",  description="Ключ"),
     *             @OA\Property(property="referral_code", type="string", description="Код пригласившего")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Успешная аутентификация или регистрация",
     *     @OA\MediaType(
     *         mediaType="text/plain",
     *         @OA\Schema(
     *             type="string",
     *             example="Success"
     *         ))
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Ошибка валидации",
     *         @OA\JsonContent(
     *             @OA\Property(property="errors", type="object", description="Список ошибок валидации")
     *         )
     *     )
     * )
     *
     * @param \Illuminate\Http\Request $request
     * @return \Illuminate\Http\JsonResponse
     */
    public function loginOrRegister(Request $request)
    {
        $request->validate([
            'phone' => 'required|string',
            'code' => 'required|integer',
            'api_key' => 'string'
        ]);
        $user = $this->phoneCodeAuthentication($request->phone, $request->code);
        if ($user) {
            if ($user->user_status === null) {
                $user->user_status = UserStatus::DocumentsNotUploaded->value;
                $user->avatar = "users/default.png";
                $user->user_type = UserType::Driver->value;
            }

            $user->code = null;
            $user->save();

            $referral = Referral::firstOrCreate(['user_id' => $user->id]);
            if (!$referral->referral_code) {
                $referral->referral_code = Str::random(20);
                if ($request->referral_code) {
                    $referral->status = ReferralStatus::Invited->name;
                    $referral->invited_id = Referral::where('referral_code', $request->referral_code)->with('user')->first()->user->id;
                } else {
                    $referral->status = ReferralStatus::NoInvited->name;
                }
                $referral->save();
            }

            if ($request->type === UserType::Driver->name) {
                $driver = Driver::firstOrCreate(['user_id' => $user->id]);
                $driverSpecification = DriverSpecification::firstOrCreate(['driver_id' => $driver->id]);
                $driverDocs = DriverDoc::firstOrCreate(['driver_id' => $driver->id]);
            }
            if ($request->type === UserType::Manager->name) {
                $manager = Manager::firstOrCreate(['user_id' => $user->id]);
                $manager->park_id = Park::where('api_key', $request->api_key)->first()->id;
                $manager->save();
            }
            $token = $user->createToken('auth_token')->plainTextToken;
            return $token;
        }
        return response()->json(null, 401);
    }

    /**
     * Выход пользователя из системы
     *
     * @OA\Post(
     *     path="api/user/logout",
     *     operationId="logout",
     *     summary="Выход пользователя из системы",
     *     tags={"Authentication"},
     *     security={{"bearerAuth": {}}},
     *     @OA\Response(
     *         response=200,
     *         description="Пользователь успешно вышел из системы",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Пользователь успешно вышел из системы")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Ошибка аутентификации",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Недопустимый токен аутентификации")
     *         )
     *     )
     * )
     *
     * @param \Illuminate\Http\Request $request
     * @return \Illuminate\Http\JsonResponse
     */
    public function logout(Request $request)
    {
        $request->user()->tokens()->delete();
        return response()->json(['message' => 'User logged out successfully']);
    }
    /**
     * Создание и отправка проверочного кода на указанный номер телефона
     *
     * @OA\Post(
     *     path="/user/code",
     *     operationId="CreateAndSendCode",
     *     summary="Создание и отправка проверочного кода на указанный номер телефона",
     *     tags={"Authentication"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="phone", type="string", example="1234567890", description="Номер телефона пользователя")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Запрос успешно выполнен",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean", example=true, description="Успешность операции")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Ошибка валидации",
     *         @OA\JsonContent(
     *             @OA\Property(property="errors", type="object", description="Список ошибок валидации")
     *         )
     *     )
     * )
     *
     * @param \Illuminate\Http\Request $request
     * @return \Illuminate\Http\JsonResponse
     */
    public function CreateAndSendCode(Request $request)
    {
        $request->validate([
            'phone' => ['required', 'string', 'max:255'],
        ]);
        $phone = $request->phone;
        $user = User::firstOrCreate(['phone' => $phone]);
        $code = rand(1000, 9999);
        // $code = 1111;
        $user->code = $code;
        if (!$user->phone) {
            $user->phone = $phone;
        }
        $user->save();
        $response = Http::get('https://sms.ru/sms/send', [
            'api_id' => 'AFA267B8-9272-4CEB-CE8B-7EE807275EA9',
            'to' => $phone,
            'msg' => 'Проверочный код: ' . $code,
            'json' => 1
        ]);
        // $response = true;
        if ($response->successful()) {
            // if ($response) {
            return response()->json(['success' => true]);
        } else {
            return response()->json(['success' => false]);
        }
    }

    private function phoneCodeAuthentication($phone, $code)
    {
        if ($code) {
            $user = User::where('phone', $phone)->where('code', $code)->first();
            if ($user) {
                Auth::login($user);
                return $user;
            }
        }
        return null;
    }

    /**
     * Удаление пользователя и связанных записей
     *
     * @OA\Delete(
     *     path="api/user",
     *     operationId="DeleteUser",
     *     summary="Удаление пользователя и связанных записей",
     *     tags={"Authentication"},
     *     security={{"bearerAuth": {}}},
     *     @OA\Response(
     *         response=200,
     *         description="Пользователь успешно удален",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Пользователь успешно удален")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Ошибка аутентификации",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Недопустимый токен аутентификации")
     *         )
     *     )
     * )
     *
     * @param \Illuminate\Http\Request $request
     * @return \Illuminate\Http\JsonResponse
     */
    public function DeleteUser(Request $request)
    {
        $user = Auth::guard('sanctum')->user();
        $driver = Driver::where('user_id', $user->id)->first();

        // Удаление папки с фотографиями пользователя
        $folderPath = 'uploads/user/' . $user->id;
        if (Storage::exists($folderPath)) {
            Storage::deleteDirectory($folderPath);
        }
        $referral = Referral::where('user_id', $user->id)->first();
        $referral->user_id = 1;
        $referral->save();
        // Удаление записей о пользователе и водителе
        $user->delete();
        if ($driver) {
            $driver->delete();
        }

        return response()->json(['message' => 'Пользователь успешно удален']);
    }
}