# garage/backend/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Мой гараж",
        "description": "Это список методов для интеграции с Мой гараж, в заголовках всех запросов должен присутствоваться API-ключ, который будет предоставлен администратором на почту клиента после регистрации клиента в сервисе администратором",
        "version": "1.0.2"
    },
    "paths": {
...
        "api/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Аутентификация пользователя или регистрация нового",
                "description": "Аутентификация пользователя или регистрация нового",
                "operationId": "loginOrRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "description": "Номер телефона пользователя",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "code": {
                                        "description": "Код аутентификации",
                                        "type": "integer"
                                    },
                                    "type": {
                                        "$ref": "#/components/schemas/UserType"
                                    },
                                    "api_key": {
                                        "description": "Ключ",
                                        "type": "string"
                                    },
                                    "referral_code": {
                                        "description": "Код пригласившего",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная аутентификация или регистрация",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Success"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Список ошибок валидации",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/user/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выход пользователя из системы",
                "description": "Выход пользователя из системы",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Пользователь успешно вышел из системы",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь успешно вышел из системы"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Недопустимый токен аутентификации"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/code": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Создание и отправка проверочного кода на указанный номер телефона",
                "description": "Создание и отправка проверочного кода на указанный номер телефона",
                "operationId": "CreateAndSendCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "description": "Номер телефона пользователя",
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запрос успешно выполнен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Успешность операции",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Список ошибок валидации",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserType": {
                "description": "The unique identifier of a product in our catalog",
                "type": "string",
                "enum": [
                    "Driver",
                    "Manager",
                    "Admin"
                ]
            }
        }
    }
}